requirements.txt
omegaconf
pydantic
python-dotenv
gradio
tqdm
soundfile

# remote режим
vllm
openai

# local режим
torch
transformers
xgrammar

# Пакеты для тестов, линтера, ..
httpx
pytest
pytest-asyncio
wemake-python-styleguide

# Пакеты для инструментов
rapidfuzz>=3.0.0